[refactor] templateColumns값 변수로 변경
